Most animations for second beta video
